+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 所有権発行時にchevreへ連携
 - 座席オファー承認処理を最適化
 - chevreでの承認検索に対応(USE_CHEVRE_SEARCH_AUTHORIZATION設定)
+- chevreでの所有権検索に対応(USE_CHEVRE_SEARCH_OWNERSHIPINFO設定)
+- 注文返品時の所有権変更をchevre連携
 
 ### Deprecated